+++ NEWS.md
@@ -1,3 +1,18 @@
+## nmarank, version 0.3-0 (2022-mm-dd)
+
+### User-visible changes
+
+* nmarank():
+  - possible values "desirable" and "undesirable" instead of "good"
+    and "bad" for argument 'small.values'
+
+### Internal changes
+
+* R package **MASS** added to imports
+
+* Do not use 'MASS::' to call ginv() from R package **MASS**
+
+
 ## nmarank, version 0.2-3 (2021-09-16)
 
 - First CRAN release